+++ python/source/state_representation/state_representation_bindings.h
@@ -23,3 +23,4 @@ void bind_state(py::module_& m);
 void bind_cartesian_space(py::module_& m);
 void bind_joint_space(py::module_& m);
 void bind_jacobian(py::module_& m);
+void bind_parameters(py::module_& m);